OpenFaaS/Functions/Auth/login/requirements.txt
Flask==2.3.2
bcrypt==4.0.1
psycopg2-binary==2.9.6
pyjwt==2.8.0